+++ src/pulsar_producer.erl
@@ -78,6 +78,7 @@
 
 -define(RECONNECT_TIMEOUT, 5_000).
 -define(LOOKUP_TOPIC_TIMEOUT, 15_000).
+-define(GET_ALIVE_PULSAR_URL, 5_000).
 
 -define(MAX_REQ_ID, 4294836225).
 -define(MAX_SEQ_ID, 18445618199572250625).
@@ -369,6 +370,7 @@ connected(state_timeout, do_connect, _State) ->
     keep_state_and_data;
 connected(state_timeout, lookup_topic_timeout, State0) ->
     log_error("timed out waiting for lookup topic response", [], State0),
+    %% todo: should demonitor reference
     State = State0#{lookup_topic_request_ref := undefined},
     ?NEXT_STATE_IDLE_RECONNECT(State);
 connected(info, ?SEND_REQ(_, _) = SendRequest, State0 = #{batch_size := BatchSize}) ->
@@ -437,7 +439,7 @@ refresh_urls_and_connect(State0) ->
      , partitiontopic := PartitionTopic
      } = State0,
     ?tp(debug, pulsar_producer_refresh_start, #{}),
-    try pulsar_client:lookup_topic_async(ClientId, PartitionTopic) of
+    try pulsar_client_manager:lookup_topic_async(ClientId, PartitionTopic) of
         {ok, LookupTopicRequestRef} ->
             State = State0#{lookup_topic_request_ref := LookupTopicRequestRef},
             {keep_state, State, [{state_timeout, ?LOOKUP_TOPIC_TIMEOUT, lookup_topic_timeout}]}
@@ -951,7 +953,7 @@ handle_lookup_topic_reply({ok, #{ proxy_through_service_url := true
     #{clientid := ClientId} = State0,
     ?tp(debug, pulsar_producer_lookup_alive_pulsar_url, #{}),
     log_debug("received topic lookup reply: ~0p", [#{proxy_through_service_url => true, broker_service_url => NewBrokerServiceURL}], State0),
-    try pulsar_client:get_alive_pulsar_url(ClientId) of
+    try pulsar_client_manager:get_alive_pulsar_url(ClientId, ?GET_ALIVE_PULSAR_URL) of
         {ok, AlivePulsarURL} ->
             maybe_connect(#{ broker_service_url => NewBrokerServiceURL
                            , alive_pulsar_url => AlivePulsarURL